requirements.txt
anthropic>=0.40.0
openai>=1.40.0
python-dotenv